backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==19.3.0
beautifulsoup4==4.11.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.6
Django==3.2.3
django-filter==23.1
django-debug-toolbar==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.1.1
gunicorn==20.1.0
Faker==12.0.1
flake8==6.0.0
flake8-isort==6.0.0
idna==2.8
importlib-metadata==1.5.0
isort==5.12.0
mccabe==0.7.0
mixer==7.1.2
more-itertools==8.2.0
psycopg2-binary==2.9.3
Pillow==9.0.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2019.3
python-dotenv==1.0.0
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.3.2.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0
